dji_command_root/requirements.txt
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
paho-mqtt==2.1.0
sqlparse==0.5.3
typing_extensions==4.15.0
django-filter
requests==2.31.0
gunicorn==21.2.0  #